Move page_titleextension to barbeque_tags
